+++ jael/lib/src/text/parser.dart
@@ -68,8 +68,10 @@ class Parser {
 
   StringLiteral implicitString() {
     if (next(TokenType.string)) {
-      return prefixParselets[TokenType.string].parse(this, _current) as StringLiteral;
-    } /*else if (next(TokenType.text)) {
+      return prefixParselets[TokenType.string].parse(this, _current)
+          as StringLiteral;
+    }
+    /*else if (next(TokenType.text)) {
 
     }*/
 
@@ -264,8 +266,7 @@ class Parser {
     if (tagName2.name != tagName.name) {
       errors.add(new JaelError(
           JaelErrorSeverity.error,
-          'Mismatched closing tags. Expected "${tagName.span
-              .text}"; got "${tagName2.name}" instead.',
+          'Mismatched closing tags. Expected "${tagName.span.text}"; got "${tagName2.name}" instead.',
           lt2.span));
       return null;
     }