+++ jael_preprocessor/lib/jael_preprocessor.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:collection';
 import 'package:file/file.dart';
 import 'package:jael/jael.dart';
+import 'package:symbol_table/symbol_table.dart';
 
 /// Modifies a Jael document.
 typedef FutureOr<Document> Patcher(Document document,
@@ -18,6 +19,7 @@ Future<Document> resolve(Document document, Directory currentDirectory,
   // Resolve all includes...
   var includesResolved =
       await resolveIncludes(document, currentDirectory, onError);
+
   var patched =
       await applyInheritance(includesResolved, currentDirectory, onError);
 
@@ -68,26 +70,63 @@ Future<Document> applyInheritance(Document document, Directory currentDirectory,
     // by filling in blocks.
     document = chain.removeFirst();
 
+    var blocks = new SymbolTable<Element>();
+
     while (chain.isNotEmpty) {
       var child = chain.removeFirst();
-      var blocks = extractBlockDeclarations(child.root, onError);
+      var scope = blocks;
+      extractBlockDeclarations(scope, child.root, onError);
       var blocksExpanded =
           await expandBlocks(document.root, blocks, currentDirectory, onError);
+
+      if (blocksExpanded == null) {
+        // This in itself is a block; expand it.
+        // TODO: Ambiguous
+
+      }
+
       document =
           new Document(child.doctype ?? document.doctype, blocksExpanded);
     }
 
     // Fill in any remaining blocks
     var blocksExpanded =
-        await expandBlocks(document.root, {}, currentDirectory, onError);
+        await expandBlocks(document.root, blocks, currentDirectory, onError);
     return new Document(document.doctype, blocksExpanded);
   }
 }
 
+List<Element> allBlocksRecursive(Element element) {
+  var out = <Element>[];
+
+  for (var e in element.children) {
+    if (e is Element && e.tagName.name == 'block') {
+      out.add(e);
+    }
+  }
+
+  for (var child in element.children) {
+    if (child is Element) {
+      var childBlocks = allBlocksRecursive(child);
+      out.addAll(childBlocks);
+//
+//      if (childBlocks.isNotEmpty) {
+//        print('<<<\n${child.span.highlight()}');
+//
+//        for (var b in childBlocks) {
+//          print('!!!!\n${b.span.highlight()}');
+//        }
+//      }
+    }
+  }
+
+  return out; //out.reversed.toList();
+}
+
 /// Extracts any `block` declarations.
-Map<String, Element> extractBlockDeclarations(
-    Element element, void onError(JaelError error)) {
-  Map<String, Element> blocks = {};
+void extractBlockDeclarations(SymbolTable<Element> blocks, Element element,
+    void onError(JaelError error)) {
+  //var blockElements = allBlocksRecursive(element);
   var blockElements =
       element.children.where((e) => e is Element && e.tagName.name == 'block');
 
@@ -104,11 +143,9 @@ Map<String, Element> extractBlockDeclarations(
           nameAttr.span));
     } else {
       var name = (nameAttr.value as StringLiteral).value;
-      blocks[name] = blockElement;
+      blocks.assign(name, blockElement);
     }
   }
-
-  return blocks;
 }
 
 /// Finds the name of the parent template.
@@ -134,8 +171,10 @@ String getParent(Document document, void onError(JaelError error)) {
 }
 
 /// Replaces any `block` tags within the element.
-Future<Element> expandBlocks(Element element, Map<String, Element> blocks,
+Future<Element> expandBlocks(Element element, SymbolTable<Element> blockss,
     Directory currentDirectory, void onError(JaelError error)) async {
+  var blocks = blockss.createChild();
+
   if (element is SelfClosingElement)
     return element;
   else if (element is RegularElement) {
@@ -166,10 +205,11 @@ Future<Element> expandBlocks(Element element, Map<String, Element> blocks,
             var name = (nameAttr.value as StringLiteral).value;
             Iterable<ElementChild> children;
 
-            if (!blocks.containsKey(name)) {
+            if (blocks.resolve(name) == null) {
+              print('Hm what? $name');
               children = child.children;
             } else {
-              children = blocks[name].children;
+              children = blocks.resolve(name).value.children;
             }
 
             expanded.addAll(children);
@@ -184,21 +224,62 @@ Future<Element> expandBlocks(Element element, Map<String, Element> blocks,
     }
 
     // Resolve all includes...
-    expanded = await Future.wait(expanded.map((c) {
-      if (c is! Element) return new Future.value(c);
-      return expandBlocks(c, blocks, currentDirectory, onError);
-    }));
+    var out = <ElementChild>[];
 
-    return new RegularElement(
+    for (var c in expanded) {
+      if (c is Element) {
+        var blocksExpanded =
+            await expandBlocks(c, blocks, currentDirectory, onError);
+        var nameAttr = c.attributes
+            .firstWhere((a) => a.name == 'name', orElse: () => null);
+        var name = (nameAttr?.value is StringLiteral)
+            ? ((nameAttr.value as StringLiteral).value)
+            : null;
+
+        if (name == null) {
+          out.add(blocksExpanded);
+        } else {
+          // This element itself resolved to a block; expand it.
+          out.addAll(blocks.resolve(name)?.value?.children ?? <ElementChild>[]);
+        }
+      } else {
+        out.add(c);
+      }
+    }
+
+    var finalElement = new RegularElement(
         element.lt,
         element.tagName,
         element.attributes,
         element.gt,
-        expanded,
+        out,
         element.lt2,
         element.slash,
         element.tagName2,
         element.gt2);
+
+    // If THIS element is a block, it needs to be expanded as well.
+    if (element.tagName.name == 'block') {
+      var nameAttr = element.attributes
+          .firstWhere((a) => a.name == 'name', orElse: () => null);
+      if (nameAttr == null) {
+        onError(new JaelError(JaelErrorSeverity.warning,
+            'Missing "name" attribute in "block" tag.', element.span));
+      } else if (nameAttr.value is! StringLiteral) {
+        onError(new JaelError(
+            JaelErrorSeverity.warning,
+            'The "name" attribute in an "block" tag must be a string literal.',
+            nameAttr.span));
+      }
+
+      var name = (nameAttr.value as StringLiteral).value;
+      print('def $name');
+      blockss.assign(name, finalElement);
+
+      return null;
+    } else {
+      return finalElement;
+    }
   } else {
     throw new UnsupportedError(
         'Unsupported element type: ${element.runtimeType}');